requirements.txt
blinker==1.8.2
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cssselect2==0.7.0
defusedxml==0.7.1
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.4.0
pycparser==2.22
requests==2.32.3
tinycss2==1.3.0
urllib3==2.2.2
webencodings==0.5.1
Werkzeug==3.0.3
